Renamed some files to correct format and updated readme description
