+++ README.md
@@ -1,3 +1,4 @@
-# MSU-Library-Web-App
+# MSU Research Guides App
 
-Lists research guides published by Mississippi State University.
+- Lists research guides published by Mississippi State University.
+- App is intended to display MSU research guides effectively to the user.
